refactor(posts): make meta column width adaptable
The meta column of the blogposts list  was fixed to 13.5rem.
I can be annoying when displaying something else but a date as it was initially planned.

Widht set to 13.5rem has been kept when displaying a date to avoid any regression.
Nevertheless, it is possible to get rid of it in the future if needed.

Feed has been adapted accordingly.